Hotfix/diff nochange foreign key

+++ diff/diff.go
@@ -413,7 +413,7 @@ func dropTableIndexes(ctx *alterCtx, dst io.Writer) (int64, error) {
 			continue
 		}
 
-		if !indexStmt.HasName() {
+		if !indexStmt.HasName() && !indexStmt.HasSymbol() {
 			return 0, errors.Errorf("can not drop index without name: %s", indexStmt.ID())
 		}
 
@@ -422,8 +422,21 @@ func dropTableIndexes(ctx *alterCtx, dst io.Writer) (int64, error) {
 		}
 		buf.WriteString("ALTER TABLE `")
 		buf.WriteString(ctx.from.Name())
-		buf.WriteString("` DROP INDEX `")
-		buf.WriteString(indexStmt.Name())
+		if indexStmt.IsForeginKey() {
+			buf.WriteString("` DROP FOREIGN KEY `")
+			if indexStmt.HasSymbol() {
+				buf.WriteString(indexStmt.Symbol())
+			} else {
+				buf.WriteString(indexStmt.Name())
+			}
+		} else {
+			buf.WriteString("` DROP INDEX `")
+			if !indexStmt.HasName() {
+				buf.WriteString(indexStmt.Symbol())
+			} else {
+				buf.WriteString(indexStmt.Name())
+			}
+		}
 		buf.WriteString("`;")
 	}
